backend/pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "0.0.0"
description = "Fastapi, devcontainer, poetry example"
authors = ["Seth Greco"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
email-validator = "^2.1.1"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
argon2-cffi = "^23.1.0"
python-jose = "^ 3.3.0"

[tool.poetry.dev-dependencies]
ruff = "^0.3.0"
pytest = "^7.4.3"
httpx = "^0.25.0"

[tool.pytest.ini_options]
python_files = "test_*.py"
testpaths = ["lib", "tests"]